requirements.txt
# 뉴스 수집
requests>=2.31
beautifulsoup4>=4.12
feedparser>=6.0
yfinance>=0.2.36

# 유틸리티
pyyaml>=6.0
